third_party/WebKit/LayoutTests/http/tests/worklet/chromium/import-filesystem-url.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Worklet: addModule() for a filesystem URL</title>
<body>
</body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

// This test should be in chromium/ because Filesystem API is implemented only
// by Chromium.
promise_test(() => {
  return new Promise(resolve => {
      const kSize = 100;
      webkitRequestFileSystem(TEMPORARY, kSize, resolve, assert_unreached);
    }).then(fs => {
      return new Promise(resolve => {
          fs.root.getFile('test.js', {create: true}, resolve, assert_unreached);
        });
    }).then(entry => {
      return paintWorklet.addModule(entry.toURL());
    }).then(() => {
      assert_unreached('addModule should fail.');
    }).catch(error => {
      console.log(error.message);
      assert_equals(error.name, 'AbortError', 'error should be an AbortError.');
    });
}, 'Importing a filesystem URL should reject the given promise with an ' +
   'AbortError.');

</script>
</body>
</html>
